backend jagir/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-dotenv==1.4.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
drf-yasg==1.21.10
fuzzywuzzy==0.18.0
hiredis==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kombu==5.5.1
Levenshtein==0.27.1
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
redis==5.2.1
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
